Working on Special Animation
Still working on the special animation shot. Rotations were not working
using reference points. Decided to have the character not move while the
special animation was going off to fix it. to have it go the other way
would require much more so for now going to go with this.